+++ NEWS.md
@@ -1,5 +1,10 @@
 # Spectra 1.15
 
+## Changes in 1.15.4
+
+- Add new `filterPeaksRanges()` function to filter mass peaks by ranges on
+  numeric spectra or peak variables.
+
 ## Changes in 1.15.3
 
 - For evaluation of the `Spectra`'s processing queue: call functions from the